ci: Lint before building
Since prettier is included as a plugin for tslint, this also checks
prettier rules.

Note about EditorConfig:

The prettier API provides a method to handling .editorconfig files[1],
but tslint-config-prettier disables this[2], so they are ignored.

If a preference is updated in one, it should also be updated in the
other.

[1]: https://prettier.io/docs/en/api.html#prettierresolveconfigfilepath-options
[2]: https://github.com/alexjoverm/tslint-config-prettier/blob/b6b1b86702f793ab214d24cdfa72cdcdefbc34c2/tools/runtime.ts#L6