ci: Allow the latest node version to fail
Currently, node 10 is in a pre-release state. It's not released yet, but
that is what `nvm install node` downloads (node 10.0.0). The 'node-sass'
doesn't support node 10 in its latest stable release, so this build
fails.

It might make sense to mandate later that builds on this version must
succeed for Travis to pass.